feat: ストーリーページに章ナビTOC＋EN/JA言語切り替えを追加
- src/lib/lang.ts: Lang型とtl()ヘルパー関数を追加
- src/lib/stories.ts: ChapterMetaにdescriptionEn（英語章説明）を追加
- story/page.tsx: 章TOC（クイックジャンプナビ）を追加、JP/ENトグル実装、
  全UIテキストを言語対応化、localStorageで言語設定を永続化
- story/[slug]/page.tsx: データ取得をServer Componentに残しUIをクライアントへ分離
- story-reader-ui.tsx: ストーリーリーダーのクライアントコンポーネント新設、
  JP/ENトグルボタン、章ラベル・前後ナビ・関連作品等を言語対応化
